put server in example
